+++ modules/terraform-zscc-asg-aws/outputs.tf
@@ -1,0 +1,4 @@
+output "availability_zone" {
+  description = "Availability zones used for ASG"
+  value       = aws_autoscaling_group.cc_asg.availability_zones
+}