+++ modules/terraform-zscc-iam-aws/variables.tf
@@ -44,3 +44,9 @@ variable "secret_name" {
   type        = string
   description = "AWS Secrets Manager Secret Name for Cloud Connector provisioning"
 }
+
+variable "asg_enabled" {
+  type        = bool
+  description = "Determines whether or not to create the cc_autoscale_lifecycle_policy IAM Policy and attach it to the CC IAM Role"
+  default     = false
+}